[Fix] mypage dispatch 개선
