+++ client/src/pages/MyPage.jsx
@@ -1,6 +1,8 @@
 ﻿import React from 'react';
 import Mypage from 'components/Mypage/Mypage';
 
-export default function MyPage() {
-  return <Mypage />;
+export default function MyPage(location) {
+
+  return <Mypage user={location.location.state.user}/>
+
 }
